Add Create account and Login features
Added features for creating and logging in with an account.

This account is created and authenticated through Firebase. Right now each account only stores a displayname in the database. Later on more features will be added for storing more data.

The login is also persisted so if a user closes and reopens the app he/she is still logged in.