Add Account activity
Added Account activity.

This activity displays the current logged in users's account. The user can see their profile image, amount of likes and display name.

The user can also edit their display name and profile image. The image picked out of the users's gallery and cropped/compressed. Then it is uploaded to Firebase.

Will have to refactor these to fragments and refactor a lot of code since right now it all is written into the activity. Will also have to edit the layout.