Implement report liking feature
When a user taps the fab in the report detail screen the report will be liked and the total amount of likes of the user will be increased by one.

This is done by using a listener who emits a complete when the request to firebase is completed and the total amount of likes is set.
The user gets an indication of having liked by the fab and a toast.