Implemented local storage
A user can edit the bottom navigation bar and toolbar's background color. This will be saved as a sharedpreference and percist in the devices storage.
Added a library that contains a colorpicker for settings menu.
Added settings menu with navigation.
Added a listener to shared preferences changes
Added some drawable files used for the colorpicker