Add fab on report detail screen to navigate to map screen
Added an extra fab on report detail screeen that enables the user to view the report on the map screen by passing the lat lang coordinates.
The map screen will zoom in on the markers location that contains the report.
If the user navigates normally to the map screen, no lat lang will be passed so the map screen will zoom in on the users current location.
Also added some more animations when navigating.